requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.76.0
distro==1.9.0
docker==7.1.0
fastapi==0.128.0
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.45.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huey==2.5.5
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mlflow==3.6.0
mlflow-skinny==3.6.0
mlflow-tracing==3.6.0
numpy==2.4.0
openai==2.8.1
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
starlette==0.50.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
Werkzeug==3.1.4
zipp==3.23.0
